requirements.txt
streamlit==1.30.0
pandas==2.1.4
plotly==5.18.0
pyyaml==6.0.1
tiktoken==0.5.2
python-dotenv==1.0.0
requests==2.31.0
regex==2023.12.25
colorama==0.4.6
pathlib==1.0.1
matplotlib==3.8.2
streamlit-echarts==0.4.0 